Se agrega la posibilidad y redondeo para poner horas y medias

+++ backend/semanal/src/main/java/com/seguimiento/semanal/entity/AdvanceDetail.java
@@ -38,5 +38,5 @@ public class AdvanceDetail {
     private String context;
 
     @Column(name = "hh")
-    private Integer hh;
+    private Double hh;
 }
